Add files via upload
This is a picture down-loader for any browser or website. Just copy the image address and paste it. Make sure to change the file format you are downloading it to. It works with videos, too, but you can't run them in VS Code.
Follow-up*: View the downloaded picture in fit-to-screen mode; otherwise, it tends to spread out.